moving taxonomy under tags
