+++ tools/sidekick/library/plugins/tags/taxonomy.js
@@ -1,4 +1,4 @@
-import ffetch from './ffetch.js';
+import ffetch from '../../../../../aemedge/scripts/ffetch.js';
 
 function titleToName(name) {
   return name.toLowerCase().replace(' ', '-');
